db2：所有插入，更新，删除方法用syn...修饰（完成）
db2：对于orderBy(String...orderBy)，参数分为两种输入情况：orderBy()和orderBy(null)。后者相当于输入了第一个String对象，值为null，然后就华丽地出错了。（完成）

db2：支持datetime类型（完成）

db2：支持外键为空的情况（完成）

db2：对于查询操作，使用PrepareStatement的占位符，一次预编译，重复修改参数值，提高效率。Hibernate也是这样的。

db2：queryMode（equal,larger,smaller,like）变为方法名（完成）

db2：完善getEntityClass()，通过父类获取子类传递的泛型（完成）

db2：queryCount返回long（完成）

DateUtil:（完成）
添加toDate(String date)方法：date可以传入yyyy-MM-dd 或 yyyy-MM-dd HH:mm:ss格式的字符串
添加toDateText(Date date)方法：返回yyyy-MM-dd格式的字符串
添加toTimeText(Date date)方法：返回HH:mm:ss格式的字符串
添加toDateTimeText(Date date)方法：返回yyyy-MM-dd HH:mm:ss格式的字符串

+++ java_lib/src/main/java/com/wang/db2/Dao.java
@@ -34,6 +34,6 @@ public interface Dao<T> {
     /**
      * 查询当前数据表的符合where条件的记录的数量
      */
-    int queryCount(Where where);
+    long queryCount(Where where);
 
 }
